code-exec-lambda/hello_world/requirements.txt
pyflakes
PyHamcrest==2.0.4
RestrictedPython==6.0
PyYAML==6.0
PyJWT>=2.6.0